refactor: del extra file StatusCode, put in utils
