+++ inc/utils.hpp
@@ -2,6 +2,7 @@
 
 #include <string>
 #include <ctime>
+#include "StatusCode.hpp"
 
 namespace utils {
 
@@ -10,5 +11,8 @@ const int timeStringBuffer = 80;
 std::string getFileExtension(const std::string& path);
 std::string getGMTString(time_t now, const char* format);
 std::string getLocaltimeString(time_t now, const char* format);
+
+std::string statusCodeToReasonPhrase(statusCode status);
+std::string getDefaultErrorPage(statusCode status);
 
 } // namespace utils
